+++ src/app/common/app.header.ts
@@ -6,7 +6,7 @@ import { checkLang } from '../service/lang.service';
 import { cssComponentMenuComponent } from '../components/css-component-menu.component';
 import { allCssComponents } from '../../framework/tools/components';
 import { renderingService } from '../service/rendering.service';
-import { gameComponentMenuComponent } from '../components/game-components/game-component-menu.component';
+import { gamePageComponent } from '../pages/game-page.component';
 
 
 export class AppHeader extends DMComponent {
@@ -115,8 +115,8 @@ export class AppHeader extends DMComponent {
       renderingService.reset();
     }
     if (router.getUrl() === 'game') {
-      gameComponentMenuComponent.createContent();
-      gameComponentMenuComponent.render();
+      gamePageComponent.createContent();
+      gamePageComponent.render();
     }
   }
